pyproject.toml
[project]
name = "text-autocomplete"
version = "0.1.0"
description = "Сравнение LSTM и DistilGPT-2 для автодополнения текста."
readme = "README.md"
requires-python = "==3.10.12"
dependencies = [
    "absl-py==2.3.0",
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.12.11",
    "aiosignal==1.3.2",
    "asttokens==3.0.0",
    "attrs==25.3.0",
    "certifi==2025.4.26",
    "charset-normalizer==3.4.2",
    "click==8.2.1",
    "colorama==0.4.6",
    "comm==0.2.2",
    "contourpy==1.3.2",
    "cycler==0.12.1",
    "datasets==3.6.0",
    "debugpy==1.8.14",
    "decorator==5.2.1",
    "dill==0.3.8",
    "evaluate==0.4.4",
    "executing==2.2.0",
    "filelock==3.18.0",
    "fonttools==4.58.2",
    "frozenlist==1.6.2",
    "fsspec==2025.3.0",
    "huggingface-hub==0.32.4",
    "idna==3.10",
    "iprogress>=0.4",
    "ipykernel>=6.0,<7.0",
    "ipython>=8.0,<9.0",
    "ipython-pygments-lexers==1.1.1",
    "jedi==0.19.2",
    "jinja2==3.1.6",
    "joblib==1.5.1",
    "jupyter-client==8.6.3",
    "jupyter-core==5.8.1",
    "kiwisolver==1.4.8",
    "markupsafe==3.0.2",
    "matplotlib==3.10.3",
    "matplotlib-inline==0.1.7",
    "mpmath==1.3.0",
    "multidict==6.4.4",
    "multiprocess==0.70.16",
    "nest-asyncio==1.6.0",
    "networkx==3.3",
    "nltk==3.9.1",
    "numpy==2.2.6",
    "packaging==25.0",
    "pandas==2.3.0",
    "parso==0.8.4",
    "pillow==11.2.1",
    "platformdirs==4.3.8",
    "prompt-toolkit==3.0.51",
    "propcache==0.3.1",
    "psutil==7.0.0",
    "pure-eval==0.2.3",
    "pyarrow==20.0.0",
    "pygments==2.19.1",
    "pyparsing==3.2.3",
    "python-dateutil==2.9.0.post0",
    "pytz==2025.2",
    "pyyaml==6.0.2",
    "pyzmq==26.4.0",
    "regex==2024.11.6",
    "requests==2.32.4",
    "rouge-score==0.1.2",
    "safetensors==0.5.3",
    "scikit-learn==1.7.0",
    "scipy==1.15.3",
    "six==1.17.0",
    "stack-data==0.6.3",
    "sympy==1.14.0",
    "threadpoolctl==3.6.0",
    "tokenizers==0.21.1",
    "torch==2.7.1",
    "tornado==6.5.1",
    "tqdm==4.67.1",
    "traitlets==5.14.3",
    "transformers==4.52.4",
    "typing-extensions==4.14.0",
    "tzdata==2025.2",
    "urllib3==2.4.0",
    "wcwidth==0.2.13",
    "xxhash==3.5.0",
    "yarl==1.20.0",
]
